+++ testsuite/src/pka.rs
@@ -10,7 +10,6 @@ use nucleo_wl55jc_bsp::hal::{
         curve::NIST_P256, EcdsaPublicKey, EcdsaSignError, EcdsaSignature, EcdsaVerifyError, Pka,
     },
     rcc,
-    util::reset_cycle_count,
 };
 use panic_probe as _;
 
@@ -18,7 +17,7 @@ const FREQ: u32 = 48_000_000;
 const CYC_PER_MICRO: u32 = FREQ / 1000 / 1000;
 
 // WARNING will wrap-around eventually, use this for relative timing only
-defmt::timestamp!("{=u32:us}", DWT::get_cycle_count() / CYC_PER_MICRO);
+defmt::timestamp!("{=u32:us}", DWT::cycle_count() / CYC_PER_MICRO);
 
 // Message hash
 const HASH: [u32; 8] = [
@@ -75,7 +74,7 @@ mod tests {
 
         cp.DCB.enable_trace();
         cp.DWT.enable_cycle_counter();
-        reset_cycle_count(&mut cp.DWT);
+        cp.DWT.set_cycle_count(0);
 
         Pka::new(dp.PKA, &mut dp.RCC)
     }