+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added a `is_pending` method to the `gpio::Exti` trait.
+- Added alarm functionality to the RTC.
 
 ## [0.3.0] - 2021-12-20
 ### Added