+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `subghz::Status`
 
 ### Removed
-- Removed `util::reset_cycle_count`, this functionality is now in `cortex-m`.
+- Removed `util::reset_cycle_count`; this functionality is now in `cortex-m`.
 
 ## [0.3.0] - 2021-12-20
 ### Added